ENH: Extracted registration result specific code for visualization and layout changes into plugin SliceTrackerRegistrationResultsPlugin
* user interface components can be hidden for different purposes
* is used by EvaluationStep and planned to be used in OverviewStep
* updated classes based on changes in base.py
* updated class diagrams
