+++ SliceTracker/CMakeLists.txt
@@ -9,6 +9,7 @@ set(MODULE_PYTHON_SCRIPTS
   SliceTrackerUtils/steps/base.py
   SliceTrackerUtils/steps/plugins/__init__.py
   SliceTrackerUtils/steps/plugins/training.py
+  SliceTrackerUtils/steps/plugins/results.py
   SliceTrackerUtils/steps/overview.py
   SliceTrackerUtils/steps/segmentation.py
   SliceTrackerUtils/steps/zFrameRegistration.py
